[windows] Partial revert #34143. Remove XFAIL.
The test started working in the last week, for unknown reasons.

Remove the XFAIL line, but keep the infra for detecting the MSVC version
for the tests. It might be useful later.

If the test starts failing again, and nobody has any idea why, the best
path forward might be marking it as UNSUPPORTED.