[IRGen] Balance out a retain count for createAsyncTask's context.
The `createAsyncTask` entry point expects the parent task to be
"guaranteed" and the function to be "owned". However, that's not easy
to model in the operand ownership map, so we consider all operands to
be "guaranteed" and balance out the reference count when lowering the builtin.

This is an egregious hack that will create a little extra reference
count traffic.