[Clang importer] Append `Async` suffix for nullable completion handler-methods.
When an Objective-C method has a nullable completion handler, that
completion handler parameter will implicitly get a `nil` default
argument. The asynchronous version and synchronous versions will
therefore have the same call signature, falling back to overload
resolution to decide amongst them. This tends to make
"fire-and-forget" uses of such APIs (which are common) harder within
asynchronous code. It also ties us to having specific async/sync
overload-resolution rules.

Append `Async` to the base name in such cases. This separates the
async from non-async cases when there will be an ambiguity, so we
don't need to lean on overload resolution.