[sil] When parsing if we find an error when parsing a SILBasicBlock emit a misc error showing the block that failed.
Otherwise, if for some reason we did not emit a diagnostic, we will just emit an
empty SILModule without any diagnostic which can be really confusing. This
ensures at least something is emitted.

I wanted to do this only if we actually emitted an error, but given that once we
emit a first error, we can't tell if we emitted more errors (we only know that
we emitted errors), we always emit this diagnostic on error. That is better than
just being confused about an empty sil module being output.