Remove side-effects semantics from unchecked_ownership_converison.
I can't reason about whether this is correct to do, because SIL both
uses the side-effects semantics to model arbitrary implicit
dependencies and simultaneously considers them to be memory
writes. What's more, the SIL instruction definitions almost *never*
specify the actual constraints that we're attempting to approximate with
the model.

Be that as it may, we can't verify SIL if we "conservatively" consider
instructions that don't write to memory to be writes. So I'll attempt
to fix the undocumented model and find out what breaks.

Fixes rdar://70737095 Unknown instruction: unchecked_ownership_conversion