[Async CC] Don't add thick placeholder to args.
Now that the convention for partial apply forwarders of async functions
is that the thick context is embedded within the async context, there is
never a need for a placeholder thick context.  Here, the placeholder
thick context is only added when the function for which a partial apply
forwarder is being emitted is not async.