More dynamic casting tests (#34732)
Add regression tests for a number of casting cases that were fixed by recent work:

* SR-13812 (aka rdar://70999129)

Verify that the new casting logic correctly handles
a generic `is` test for a struct conforming to a protocol.

* SR-8964 (aka rdar://45217461)

Test casts from Any! holding Error types

* SR-1999 (aka rdar://48769924)

* SR-6279 (aka rdar://35321438

* rdar://59844232

* SR-12025

Note that SR-12025 is not actually fixed yet; this test is marked "expected failure" until it is.