[windows][msvc2017] Avoid initializer list constructor ambiguity.
MSVC2017 seems to not fallback to the correct overloaded constructor
when an initializer list constructor seems to be used, but only fails
because the conversion rules. Using parenthesis instead of braces seems
to indicate MSVC2017 the right constructor to use, and should work in
the rest of the compilers as well.

Seems that MSVC2019 is more resilient to this ambiguity.

Introduced in #34808 and started failing https://ci-external.swift.org/job/oss-swift-windows-x86_64/5953/