[Async CC] Never map returns to native explosions.
Previously, when lowering the entry point of an async function, the
returned values were lowered to explosions that matched those of sync
functions, namely native explosions. That is incorrect for async
functions where the structured values are within the async context.

Here, that error is fixed, by loading the values returned from the call
out of the async context passed in.

rdar://problem/71641793