[sil] Remove usage from TypeLowering of SILBuilder::create*AndFold().
These create*AndFold APIs are actively harmful when used in TypeLowering since:

1. In general they are a problem since it is weird for a builder API to remove
   an instruction.

2. These APIs do not take an erase callback that must be used in passes that
   need to update state before erasing the instruction.

3. The typelowering APIs that use this are emitDestroyValue/etc which are the
   main APIs that we are using to write code that works with OSSA/non-OSSA
   SIL. So we are going to use these APIs in many more places, introducing this
   bug in many places.

With that in mind, I have been committing small cheap ARC optimizations to
GuaranteedARCOpts (SemanticARCOpts with expensive optimizations turned off) so
that we can eliminate this without massively churning the code. We are at this
stage now, so it makes sense to turn this off.