[ast] Add LLVM options to the diagnostic verifier that causes it to assert when it emits various failure diagnostics.
Sometimes it is hard to determine just by reading the output why the verifier
failed. This patch adds a compiler dev LLVM option that causes the verifier to
assert in such a situation allowing one to drop immediately into the debugger
and figure out what the verifier was doing. IOW improve debuggability of the
diagnostic verifier.