[diagnostic-verifier] Shrink DiagnosticVerifier::verifyFile by splitting out parsing of expected diagnostics into a helper function.
This function is really large and is doing a bunch of different things. Its size
makes it hard to understand. This is a first commit in a series of commits that
splits it up into several local functions to ease readability. I am doing these
commits using Xcode's refactoring and then git-clang-formatting.

As an aside: the reason why I am doing this is that I want to add some compiler
developer only LLVM options that cause the diagnostic verifier to assert when it
finds certain errors. This then makes it easier to debug why the error occurred
from within the implementation since one can just run it in the debugger and
immediately hit the issue. This is following the prior art where we have similar
compiler developer options.